requirements.txt
pandas
requests
pyyaml
jsonschema
jsonref